requirements.txt
#

# These requirements were autogenerated by pipenv

# To regenerate from the project's Pipfile, run:

#

#    pipenv lock --requirements

#

# To regenerate virtual env from this requirements.txt file, run:

#

#      pipenv install -r requirements.txt



#   pipenv install drf-yasg

django==3.2.8

djangorestframework==3.12.4

djangorestframework-simplejwt==5.0.0

django-cors-headers==3.10.0

django-filter==21.1

django-multiselectfield==0.1.12

django-simple-history==3.0.0

pillow==8.4.0

psycopg2==2.9.1

pyyaml==6.0

uritemplate==4.1.1

beautifulsoup4==4.10.0

drf-yasg==1.20.0